requirements.txt
asgiref==3.2.7
Django==3.0.7
django-stdimage==5.1.1
Pillow==7.1.2
progressbar2==3.51.3
python-utils==2.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
